ingestao-infoprice/requirements.txt
pandas
google-cloud-storage
google-cloud-bigquery
load_dotenv
pyarrow